Move PopulateThetaMatrixFromCacheEntry into CacheManager
