+++ src/artm/core/cache_manager.h
@@ -19,6 +19,12 @@
 namespace artm {
 namespace core {
 
+// CacheManager class is responsible for caching ThetaMatrix in between calls to different APIs.
+// This class is used when the user calls FitOffline / FitOnline / Transfor to store the resulting theta matrix.
+// (at least when theta_matrix_type is set to ThetaMatrixType_Cache).
+// Later user may retrieve the data from CacheManager via calls to ArtmRequestThetaMatrix.
+// The cache is organized as a set of entries, each entry associated with a single batch.
+// The key in the cache corresponds to 'batch.id' field.
 class CacheManager : boost::noncopyable {
  public:
   CacheManager();
