+++ src/artm/core/helpers.cc
@@ -223,128 +223,5 @@ void BatchHelpers::SaveMessage(const std::string& full_filename,
   fout.close();
 }
 
-// ToDo(sashafrey): this method has grown too big and complicated.
-// It needs to be refactored.
-bool BatchHelpers::PopulateThetaMatrixFromCacheEntry(
-    const DataLoaderCacheEntry& cache,
-    const GetThetaMatrixArgs& get_theta_args,
-    ::artm::ThetaMatrix* theta_matrix) {
-  if (get_theta_args.topic_index_size() != 0 && get_theta_args.topic_name_size() != 0)
-    BOOST_THROW_EXCEPTION(InvalidOperation(
-    "GetThetaMatrixArgs.topic_name and GetThetaMatrixArgs.topic_index must not be used together"));
-
-  auto& args_topic_name = get_theta_args.topic_name();
-  auto& args_topic_index = get_theta_args.topic_index();
-  const bool has_sparse_format = get_theta_args.matrix_layout() == GetThetaMatrixArgs_MatrixLayout_Sparse;
-  const bool sparse_cache = cache.topic_index_size() > 0;
-  bool use_all_topics = false;
-
-  std::vector<int> topics_to_use;
-  if (args_topic_index.size() > 0) {
-    for (int i = 0; i < args_topic_index.size(); ++i) {
-      int topic_index = args_topic_index.Get(i);
-      if (topic_index < 0 || topic_index >= cache.topic_name_size()) {
-        std::stringstream ss;
-        ss << "GetThetaMatrixArgs.topic_index[" << i << "] == " << topic_index << " is out of range.";
-        BOOST_THROW_EXCEPTION(artm::core::InvalidOperation(ss.str()));
-      }
-      topics_to_use.push_back(topic_index);
-    }
-  } else if (args_topic_name.size() != 0) {
-    for (int i = 0; i < args_topic_name.size(); ++i) {
-      int topic_index = repeated_field_index_of(cache.topic_name(), args_topic_name.Get(i));
-      if (topic_index == -1) {
-        std::stringstream ss;
-        ss << "GetThetaMatrixArgs.topic_name[" << i << "] == " << args_topic_name.Get(i)
-           << " does not exist in MasterModelConfig.topic_name";
-        BOOST_THROW_EXCEPTION(artm::core::InvalidOperation(ss.str()));
-      }
-
-      assert(topic_index >= 0 && topic_index < cache.topic_name_size());
-      topics_to_use.push_back(topic_index);
-    }
-  } else {  // use all topics
-    assert(cache.topic_name_size() > 0);
-    for (int i = 0; i < cache.topic_name_size(); ++i)
-      topics_to_use.push_back(i);
-    use_all_topics = true;
-  }
-
-  // Populate topics_count and topic_name fields in the resulting message
-  ::google::protobuf::RepeatedPtrField< ::std::string> result_topic_name;
-  for (int topic_index : topics_to_use)
-    result_topic_name.Add()->assign(cache.topic_name(topic_index));
-
-  if (theta_matrix->topic_name_size() == 0) {
-    // Assign
-    theta_matrix->set_topics_count(result_topic_name.size());
-    assert(theta_matrix->topic_name_size() == 0);
-    for (const TopicName& topic_name : result_topic_name)
-      theta_matrix->add_topic_name(topic_name);
-  } else {
-    // Verify
-    if (theta_matrix->topics_count() != result_topic_name.size())
-      BOOST_THROW_EXCEPTION(artm::core::InternalError("theta_matrix->topics_count() != result_topic_name.size()"));
-    for (int i = 0; i < theta_matrix->topic_name_size(); ++i) {
-      if (theta_matrix->topic_name(i) != result_topic_name.Get(i))
-        BOOST_THROW_EXCEPTION(artm::core::InternalError("theta_matrix->topic_name(i) != result_topic_name.Get(i)"));
-    }
-  }
-
-  bool has_title = (cache.item_title_size() == cache.item_id_size());
-  for (int item_index = 0; item_index < cache.item_id_size(); ++item_index) {
-    theta_matrix->add_item_id(cache.item_id(item_index));
-    if (has_title) theta_matrix->add_item_title(cache.item_title(item_index));
-    ::artm::FloatArray* theta_vec = theta_matrix->add_item_weights();
-
-    const artm::FloatArray& item_theta = cache.theta(item_index);
-    if (!has_sparse_format) {
-      if (sparse_cache) {
-        // dense output -- sparse cache
-        for (unsigned index = 0; index < topics_to_use.size(); ++index) {
-          int topic_index = repeated_field_index_of(cache.topic_index(item_index).value(), topics_to_use[index]);
-          theta_vec->add_value(topic_index != -1 ? item_theta.value(topic_index) : 0.0f);
-        }
-      } else {
-        // dense output -- dense cache
-        for (int topic_index : topics_to_use)
-          theta_vec->add_value(item_theta.value(topic_index));
-      }
-    } else {
-      ::artm::IntArray* sparse_topic_index = theta_matrix->add_topic_index();
-      if (sparse_cache) {
-        // sparse output -- sparse cache
-        for (int index = 0; index < cache.topic_index(item_index).value_size(); ++index) {
-          int topic_index = cache.topic_index(item_index).value(index);
-          if (use_all_topics) {
-            theta_vec->add_value(item_theta.value(index));
-            sparse_topic_index->add_value(topic_index);
-          } else {
-            for (unsigned i = 0; i < topics_to_use.size(); ++i) {
-              if (topics_to_use[i] == topic_index) {
-                theta_vec->add_value(item_theta.value(index));
-                sparse_topic_index->add_value(topic_index);
-                break;
-              }
-            }
-          }
-        }
-      } else {
-        // sparse output -- dense cache
-        for (unsigned index = 0; index < topics_to_use.size(); index++) {
-          int topic_index = topics_to_use[index];
-          float value = item_theta.value(topic_index);
-          if (value >= get_theta_args.eps()) {
-            theta_vec->add_value(value);
-            sparse_topic_index->add_value(index);
-          }
-        }
-      }
-    }
-  }
-
-  return true;
-}
-
 }  // namespace core
 }  // namespace artm
